refactor(vmm): Move KvmCpuidFlags and remove cpuid_ffi.rs
